stage 1

+++ interpolater.cpp
@@ -23,11 +23,11 @@ std::vector<double> Interpolater::Newton(const std::vector<double>& xs,
 	const std::vector<double>& ys, const std::vector<double>& x0) {
 	double* f = getDiffTable(xs, ys);
 	std::vector<double> res;
-	for (int k = 0; k < x0.size(); k++) {
+    for (size_t k = 0; k < x0.size(); k++) {
 		double sum = 0;
-		for (int i = 0; i < xs.size(); i++) {
+        for (size_t i = 0; i < xs.size(); i++) {
 			double mul = f[i];
-			for (int j = 0; j < i; j++) {
+            for (size_t j = 0; j < i; j++) {
 				mul *= x0[k] - xs[j];
 			}
 			sum += mul;
@@ -42,7 +42,7 @@ std::vector<double> Interpolater::Newton(const std::vector<double>& xs,
 double Interpolater::Lagrange(const std::vector<double>& xs, 
 	const std::vector<double>& ys, double x) {
 
-	int n = xs.size();
+    size_t n = xs.size();
 	double res = 0;
 	for (size_t i = 0; i < n; i++) {
 		double tmp = 1.0;
@@ -62,7 +62,7 @@ double Interpolater::Lagrange(const std::vector<double>& xs,
 std::vector<double> Interpolater::Lagrange(const std::vector<double>& xs,
 	const std::vector<double>& ys, const std::vector<double>& x0) {
 	std::vector<double> res;
-	for (int i = 0; i < x0.size(); i++) {
+    for (size_t i = 0; i < x0.size(); i++) {
 		res.push_back(Interpolater::Lagrange(xs, ys, x0[i]));
 	}
 	return res;
